further fixes to linting
