+++ _includes/footer_scripts.html
@@ -18,4 +18,4 @@
   </script>
 
   <script src="{{ site.baseurl }}/js/vendor/uswds.v0.9.6.js"></script>
-  <script src="{{ site.baseurl }}/js/index.js"></script>
+  <script src="{{ site.baseurl }}/assets/bundle.js"></script>
